extract plugin loading to new plugin manager class
